Dev heroku config

+++ temperature/models.py
@@ -3,7 +3,7 @@ from django.db import models
 # Create your models here.
 
 class Temperature(models.Model):
-    temperature = models.SmallIntegerField()
+    temperature = models.FloatField()
     date = models.CharField(max_length=20)
     time = models.CharField(max_length=20)
